test_get_pet_by_status parametrized
